third_party/WebKit/LayoutTests/http/tests/security/referrerPolicyHeader/worker-with-header.html

<!DOCTYPE html>
<title>Worker served with Referrer-Policy header</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js?pipe=sub"></script>
<script>
async_test(function(t) {
    var worker = new Worker('resources/worker.php');
    worker.onmessage = t.step_func(function (e) {
      assert_equals(e.data, get_host_info()['HTTP_ORIGIN'] + '/');
      t.done();
    });
  }, 'Worker script served with a Referrer-Policy header');
</script>
